requirement.txt
alembic==1.7.5
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
psycopg2-binary==2.9.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==5.4.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
